requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
Flask-Cors==5.0.0
SQLAlchemy==2.0.36
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.1
alembic==1.13.2
